Implementação da regra de usuario duplicado no cadastro

+++ Scripts/usuarios.js
@@ -16,7 +16,22 @@ function cadastrarUsuario(nome, email, senha) {
 
     var usuarios = buscarUsuarios();
 
+    if (usuarioJaExiste(usuario, usuarios)) {
+        alert("Já existe um usuário cadastrado com o e-mail informado");
+        return false;
+    }
+
     usuarios.push(usuario);
 
     localStorage.setItem(chaveUsuarios, JSON.stringify(usuarios));
+
+    return true;
+}
+
+function usuarioJaExiste(usuario, usuarios) {
+    var usuarioEncontrado = $.grep(usuarios, function(item, indice) {
+        return item.email.trim() === usuario.email.trim();
+    });
+
+    return usuarioEncontrado.length > 0;
 }
